Rescope integration helpers around runtime adapters
