+++ src/dc43/components/contract_store/__init__.py
@@ -1,13 +1,23 @@
-"""Contract store interfaces and reference implementations."""
+"""Contract store interfaces, gateways, and reference implementations."""
 
 from .interface import ContractStore
 from .impl.filesystem import FSContractStore
 from .impl.delta import DeltaContractStore
 from .impl.collibra import CollibraContractStore
+from .collibra_gateway import (
+    CollibraContractGateway,
+    ContractSummary,
+    HttpCollibraContractGateway,
+    StubCollibraContractGateway,
+)
 
 __all__ = [
     "ContractStore",
     "FSContractStore",
     "DeltaContractStore",
     "CollibraContractStore",
+    "CollibraContractGateway",
+    "ContractSummary",
+    "HttpCollibraContractGateway",
+    "StubCollibraContractGateway",
 ]
