+++ src/dc43/components/data_quality/__init__.py
@@ -6,6 +6,7 @@ from .engine import (
     compute_metrics,
     expectations_from_contract,
 )
+from .validation import ValidationResult, apply_contract, validate_dataframe
 
 __all__ = [
     "DQClient",
@@ -13,4 +14,7 @@ __all__ = [
     "attach_failed_expectations",
     "compute_metrics",
     "expectations_from_contract",
+    "ValidationResult",
+    "validate_dataframe",
+    "apply_contract",
 ]
