+++ src/dc43/components/integration/spark_io.py
@@ -16,7 +16,7 @@ from dc43.components.contract_drafter import draft_from_dataframe
 from dc43.components.contract_store import ContractStore
 from dc43.components.data_quality import DQClient, DQStatus
 from dc43.components.data_quality.engine import compute_metrics
-from dc43.components.integration.validation import (
+from dc43.components.data_quality.validation import (
     ValidationResult,
     apply_contract,
     validate_dataframe,
@@ -25,7 +25,35 @@ from dc43.odcs import contract_identity, ensure_version
 from dc43.versioning import SemVer
 from open_data_contract_standard.model import OpenDataContractStandard, Server  # type: ignore
 
-from .dataset import dataset_id_from_ref, get_delta_version
+
+def get_delta_version(
+    spark: SparkSession,
+    *,
+    table: Optional[str] = None,
+    path: Optional[str] = None,
+) -> Optional[str]:
+    """Return the latest Delta table version as a string if available."""
+
+    try:
+        ref = table if table else f"delta.`{path}`"
+        row = spark.sql(f"DESCRIBE HISTORY {ref} LIMIT 1").head(1)
+        if not row:
+            return None
+        # versions column name can be 'version'
+        v = row[0][0]
+        return str(v)
+    except Exception:
+        return None
+
+
+def dataset_id_from_ref(*, table: Optional[str] = None, path: Optional[str] = None) -> str:
+    """Build a dataset id from a table name or path (``table:...``/``path:...``)."""
+
+    if table:
+        return f"table:{table}"
+    if path:
+        return f"path:{path}"
+    return "unknown"
 
 logger = logging.getLogger(__name__)
